Do not setWindowStateInternal in QWaylandWindow's constructor
Do not call setWindowStateInternal since we have our own window state
logic in webOS and want it to be used always. Instead, the identical
method defined in the descendant class will be called in its
constructor. QWaylandWindow::setWindowStateInternal may trigger an
action for wl_shell_surface which is not desirable in webOS.

Change-Id: Ia3e5f4133784c965440a42fe9fc9eca99badcfa3